Restful done

+++ src/main/java/com/cg/sprint1_onlineplantnursery/controller/SeedController.java
@@ -2,6 +2,7 @@ package com.cg.sprint1_onlineplantnursery.controller;
 
 import java.util.List;
 import java.util.Optional;
+import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,10 +16,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
 import com.cg.sprint1_onlineplantnursery.entity.Seed;
+import com.cg.sprint1_onlineplantnursery.exception.SeedIdNotFoundException;
 import com.cg.sprint1_onlineplantnursery.service.ISeedService;
-import com.cg.sprint1_onlineplantnursery.service.ISeedServiceImpl;
 
 @RestController
 @RequestMapping("/rest")
@@ -27,51 +27,70 @@ public class SeedController extends WebSecurityConfigurerAdapter{
 	@Autowired
 	private ISeedService seedService;
 	
-	 @Override
-	    protected void configure(HttpSecurity http) throws Exception {
-	       super.configure(http);
-	       http.csrf().disable();
-	    }
-	
+	@Override
+	protected void configure(HttpSecurity http) throws Exception {
+	    super.configure(http);
+	    http.csrf().disable();
+	}
+	 
 	@PostMapping("/seeds")
-	public ResponseEntity<Seed> addSeed(@RequestBody Seed seed){
+	public ResponseEntity<Seed> addSeed(@Valid @RequestBody Seed seed){
 		seedService.addSeed(seed);
 		return new ResponseEntity<Seed>(seed,HttpStatus.CREATED);
 	}
 	
 	@PutMapping("/seeds")
 	public ResponseEntity<Seed> updateSeed(@RequestBody Seed seed){
-		seedService.updateSeed(seed);
-		return new ResponseEntity<Seed>(seed,HttpStatus.OK);
+		if(seedService.getSeed(seed.getId()).isPresent())
+		{
+			seedService.updateSeed(seed);
+			return new ResponseEntity<Seed>(seed,HttpStatus.CREATED);
+		}
+		else 
+		{
+			return new ResponseEntity<Seed>(seed,HttpStatus.BAD_REQUEST);
+		}
 	}
 	
 	@DeleteMapping("/seeds")
-	public ResponseEntity<Seed> deleteSeed(@RequestBody Seed seed){
-		Seed seedDeleted = seedService.deleteSeed(seed);
-		return new ResponseEntity<Seed>(seedDeleted,HttpStatus.OK);
+	public ResponseEntity<Seed> deleteSeed(@RequestBody Seed seed) throws SeedIdNotFoundException{
+		if(seedService.getSeed(seed.getId()).isPresent())
+		{
+			Seed seedDeleted = seedService.deleteSeed(seed);
+			return new ResponseEntity<Seed> (seedDeleted,HttpStatus.ACCEPTED);
+		}
+		else {
+			throw new SeedIdNotFoundException("Cannot be deleted...(Invalid seed Id)");
+		}	
 	}
-	
+		
 	@GetMapping("/seeds/id/{id}")
-	public ResponseEntity<Optional<Seed>> viewSeed(@PathVariable int id){
-		Optional<Seed> seedRes = seedService.viewSeed(id);
-		return new ResponseEntity<Optional<Seed>>(seedRes,HttpStatus.OK);
+	public ResponseEntity<Optional<Seed>> viewSeed(@PathVariable int id) throws SeedIdNotFoundException{
+		if(seedService.getSeed(id).isPresent()) {
+			Optional<Seed> seedRes = seedService.getSeed(id);
+			return new ResponseEntity<Optional<Seed>>(seedRes,HttpStatus.OK);
+		}
+		else {
+			throw new SeedIdNotFoundException("Invalid Seed ID...Cannot fetch entities ");
+		}
+		
 	}
 	
 	@GetMapping("/seeds/commonName/{commonName}")
-	public ResponseEntity<Seed> viewSeed(@PathVariable String commonName){
-		Seed seedRes = seedService.viewSeed(commonName);
-		return new ResponseEntity<Seed>(seedRes,HttpStatus.ACCEPTED);
+	public ResponseEntity<Seed> viewSeed(@PathVariable String commonName) throws SeedIdNotFoundException{
+			Seed seedRes = seedService.getSeed(commonName);
+			return new ResponseEntity<Seed>(seedRes,HttpStatus.ACCEPTED);	
 	}
 	
 	@GetMapping("/seeds")
 	public ResponseEntity<List<Seed>> viewAllSeed(){
-		List<Seed> seedList = seedService.viewAllSeeds();
+		List<Seed> seedList = seedService.getSeeds();
 		return new ResponseEntity<List<Seed>>(seedList,HttpStatus.ACCEPTED);
 	}
 	
 	@GetMapping("/seeds/typeOfSeed/{typeOfSeed}")
 	public ResponseEntity<List<Seed>> viewAllSeed(@PathVariable String typeOfSeed){
-		List<Seed> seedList = seedService.viewAllSeeds(typeOfSeed);
+		List<Seed> seedList = seedService.getSeeds(typeOfSeed);
 		return new ResponseEntity<List<Seed>>(seedList,HttpStatus.ACCEPTED);
 	}
 }
